% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{.validateJSON}
\alias{.validateJSON}
\title{validateJSON}
\usage{
.validateJSON(json)
}
\arguments{
\item{json}{list. Answer json from SolarDB}
}
\description{
validateJSON
}
\examples{
\dontrun{
.getJSON()
}

}
\author{
Mathieu Delsaut, \email{mathieu.delsaut@univ-reunion.fr}
}
