% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\name{getData}
\alias{getData}
\title{Get data from SolarDB}
\usage{
getData(sites = NULL, types = NULL, sensors = NULL, start = NULL, stop = NULL)
}
\description{
This function allow query the data from SolarDB.
}
\examples{
\dontrun{
 sites()
}

}
\author{
Mathieu Delsaut, \email{mathieu.delsaut@univ-reunion.fr}
}
