% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deprecated.R
\name{getDataCountByDay}
\alias{getDataCountByDay}
\title{Get daily data count for a sensor from SolarDB (Deprecated)}
\usage{
getDataCountByDay(site, sensor)
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#deprecated}{\figure{lifecycle-deprecated.svg}{options: alt='[Deprecated]'}}}{\strong{[Deprecated]}}

This function is deprecated. Please use \code{sdb_daily_counts()} instead.
}
\keyword{internal}
