% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\name{getXtsData}
\alias{getXtsData}
\title{Get data from SolarDB into Xts format}
\usage{
getXtsData(
  sites = NULL,
  types = NULL,
  sensors = NULL,
  start = NULL,
  stop = NULL
)
}
\arguments{
\item{sites}{character (or character vector). List of sites}

\item{types}{character (or character vector). List of types}

\item{sensors}{character (or character vector). List of sensors}

\item{start}{character. Date, Timestamp or duration}

\item{stop}{character. Date, Timestomp or duration}
}
\description{
This function allow query the data from SolarDB and return the result into
xts format.
}
\examples{
\dontrun{
 getXtsData()
}

}
\author{
Mathieu Delsaut, \email{mathieu.delsaut@univ-reunion.fr}
}
