% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metadata.R
\name{instruments}
\alias{instruments}
\title{Instruments from SolarDB}
\usage{
instruments(id = NULL, name = NULL, label = NULL, serial = NULL)
}
\arguments{
\item{id}{character. The _id from the table instruments}

\item{name}{character. The name from the table instruments}

\item{label}{character. The label from the table instruments}

\item{serial}{character. The serial from the table instruments}
}
\description{
This function allow know the instruments from SolarDB.
}
\examples{
\dontrun{
 instruments()
}

}
\author{
Mathieu Delsaut, \email{mathieu.delsaut@univ-reunion.fr}
}
