% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/login.R
\name{login}
\alias{login}
\title{Login to SolarDB with token}
\usage{
login(token = NULL)
}
\description{
This function allow to login to SolarDB and store the cookie needed to
authenticate user.
}
\examples{
\dontrun{
 login()
}

}
\author{
Mathieu Delsaut, \email{mathieu.delsaut@univ-reunion.fr}
}
