% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deprecated.R
\name{logout}
\alias{logout}
\title{Logout from SolarDB (Deprecated)}
\usage{
logout()
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#deprecated}{\figure{lifecycle-deprecated.svg}{options: alt='[Deprecated]'}}}{\strong{[Deprecated]}}

This function is deprecated. Please use \code{sdb_logout()} instead.
}
\keyword{internal}
