% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metadata.R
\name{measures}
\alias{measures}
\title{Measures from SolarDB}
\usage{
measures(id = NULL, names = NULL, type = NULL, nested = FALSE)
}
\description{
This function allow know the measures from SolarDB.
}
\examples{
\dontrun{
 measures()
}

}
\author{
Mathieu Delsaut, \email{mathieu.delsaut@univ-reunion.fr}
}
