% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shiny.R
\name{missing}
\alias{missing}
\title{Launch missing app in the default browser}
\usage{
missing()
}
\description{
Shiny App to visualise missing data from SolarDB
}
\examples{
if (interactive()) {
  missing()
}
}
\author{
Mathieu Delsaut, \email{mathieu.delsaut@univ-reunion.fr}
}
