% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shiny.R
\name{missings}
\alias{missings}
\title{Launch missings app in the default browser}
\usage{
missings()
}
\description{
Shiny App to visualise missing data from SolarDB
}
\examples{
if (interactive()) {
  missings()
}
}
\author{
Mathieu Delsaut, \email{mathieu.delsaut@univ-reunion.fr}
}
