% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metadata.R
\name{models}
\alias{models}
\title{Models from SolarDB}
\usage{
models(id = NULL, name = NULL, type = NULL)
}
\arguments{
\item{id}{character. The _id from the table models}

\item{name}{character. The name from the table models}

\item{type}{character. The type from the table models}
}
\description{
This function allow know the models from SolarDB.
}
\examples{
\dontrun{
 models()
}

}
\author{
Mathieu Delsaut, \email{mathieu.delsaut@univ-reunion.fr}
}
