% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rSolarDB-package.R
\docType{package}
\name{rSolarDB-package}
\alias{rSolarDB}
\alias{rSolarDB-package}
\title{rSolarDB: r library to access SolarDB database}
\description{
Package allow to download data and metadata from SolarDB. For more information visit \url{https://galilee.univ-reunion.fr}.
}
\author{
\strong{Maintainer}: Mathieu Delsaut \email{mathieu.delsaut@univ-reunion.fr} (\href{https://orcid.org/0000-0003-3878-1642}{ORCID})

}
\keyword{internal}
