% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metadata.R
\name{sdb_campaigns}
\alias{sdb_campaigns}
\title{Get Campaign Information from SolarDB}
\usage{
sdb_campaigns(id = NULL, name = NULL, territory = NULL, alias = NULL)
}
\arguments{
\item{id}{character. Campaign identifier}

\item{name}{character. Campaign name}

\item{territory}{character. Geographic territory of the campaign}

\item{alias}{character. Alternative site name}
}
\value{
list. Campaign information if successful, NULL otherwise
}
\description{
Retrieves information about measurement campaigns from the SolarDB database.
Allows filtering by various campaign attributes.
}
\examples{
# Get all campaigns
all_campaigns <- sdb_campaigns()

# Get campaign by name
marla_campaign <- sdb_campaigns(name = "marla")

# Filter by territory
reunion_campaigns <- sdb_campaigns(territory = "Reunion")

}
\author{
Mathieu Delsaut, \email{mathieu.delsaut@univ-reunion.fr}
}
