% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\name{sdb_daily_counts}
\alias{sdb_daily_counts}
\title{Get daily data count for a sensor from SolarDB}
\usage{
sdb_daily_counts(site, sensor)
}
\arguments{
\item{site}{character. Site alias where the sensor is located}

\item{sensor}{character. Sensor ID to get counts for}
}
\value{
A list containing daily measurement counts for the specified sensor
}
\description{
Retrieves the count of measurements per day for a specific sensor. This is useful
for checking data availability and completeness.
}
\examples{
# Get daily counts for a GHI sensor at MOUFIA
counts <- sdb_daily_counts("marla", "DHI_qn01_Avg")

}
\author{
Mathieu Delsaut, \email{mathieu.delsaut@univ-reunion.fr}
}
