% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\name{sdb_data}
\alias{sdb_data}
\title{Query time series data from SolarDB}
\usage{
sdb_data(
  sites = NULL,
  types = NULL,
  sensors = NULL,
  start = NULL,
  stop = NULL,
  aggrFn = NULL,
  aggrEvery = NULL,
  format = "list"
)
}
\arguments{
\item{sites}{character vector. List of sites to query}

\item{types}{character vector. List of data types to query}

\item{sensors}{character vector. List of sensor IDs to query}

\item{start}{character. Start date/time in ISO format or duration (e.g. "-1d" for last day)}

\item{stop}{character. End date/time in ISO format or duration}

\item{aggrFn}{character. Aggregation function (e.g. "mean", "sum", "count")}

\item{aggrEvery}{character. Aggregation period (e.g. "1h", "1d")}

\item{format}{character. Format of the returned data (e.g. "list" (default), "xts", "data.table")}
}
\value{
A nested list of time series data organized by site and sensor
}
\description{
Retrieves time series data from the SolarDB API with flexible filtering and aggregation options.
}
\examples{
# Get raw data for last day
sdb_data(sites = "marla", types = "GHI", start = "-1d")

# Get hourly averages for a date range
sdb_data(sensors = "GHI", start = "2023-01-01", stop = "2023-01-31",
  aggrFn = "mean", aggrEvery = "1h")

}
\author{
Mathieu Delsaut, \email{mathieu.delsaut@univ-reunion.fr}
}
