% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metadata.R
\name{sdb_instruments}
\alias{sdb_instruments}
\title{Get Instrument Information from SolarDB}
\usage{
sdb_instruments(id = NULL, name = NULL, label = NULL, serial = NULL)
}
\arguments{
\item{id}{character. Instrument identifier}

\item{name}{character. Instrument name/model}

\item{label}{character. Display label for the instrument}

\item{serial}{character. Serial number of the instrument}
}
\value{
list. Instrument information if successful, NULL otherwise
}
\description{
Retrieves information about measurement instruments from the SolarDB database.
Allows filtering by various instrument attributes.
}
\examples{
# Get all instruments
all_instruments <- sdb_instruments()

# Get instrument by serial number
specific_instrument <- sdb_instruments(serial = "SN123456")

}
\author{
Mathieu Delsaut, \email{mathieu.delsaut@univ-reunion.fr}
}
