% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/login.R
\name{sdb_login}
\alias{sdb_login}
\title{Login to SolarDB with token}
\usage{
sdb_login(token = NULL)
}
\arguments{
\item{token}{character. The token to connect to SolarDB}
}
\description{
This function allow to login to SolarDB and store the cookie needed to
authenticate user.
}
\examples{
\dontrun{
 sdb_login()
}

}
\author{
Mathieu Delsaut, \email{mathieu.delsaut@univ-reunion.fr}
}
