% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/login.R
\name{sdb_logout}
\alias{sdb_logout}
\title{Logout to SolarDB}
\usage{
sdb_logout()
}
\description{
This function allow to logout to SolarDB.
}
\examples{
\dontrun{
 sdb_logout()
}

}
\author{
Mathieu Delsaut, \email{mathieu.delsaut@univ-reunion.fr}
}
