% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metadata.R
\name{sdb_models}
\alias{sdb_models}
\title{Get Model Information from SolarDB}
\usage{
sdb_models(id = NULL, name = NULL, type = NULL)
}
\arguments{
\item{id}{character. Model identifier}

\item{name}{character. Model name}

\item{type}{character. Model type/category}
}
\value{
list. Model information if successful, NULL otherwise
}
\description{
Retrieves information about instrument models from the SolarDB database.
Allows filtering by various model attributes.
}
\examples{
# Get all models
all_models <- sdb_models()

# Get models of specific type
sensor_models <- sdb_models(type = "sensor")

}
\author{
Mathieu Delsaut, \email{mathieu.delsaut@univ-reunion.fr}
}
