% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\name{sdb_sensors}
\alias{sdb_sensors}
\title{Get list of sensor IDs from SolarDB}
\usage{
sdb_sensors(sites = NULL, types = NULL)
}
\arguments{
\item{sites}{character vector. List of site aliases to filter sensors}

\item{types}{character vector. List of data types to filter sensors}
}
\value{
A list of sensor IDs matching the filter criteria
}
\description{
Retrieves the list of sensor IDs from the SolarDB API, optionally filtered by sites and types.
}
\examples{
all_sensors <- sdb_sensors()
head(all_sensors)

marla_sensors <- sdb_sensors(sites = "marla")
marla_sensors

ghi_sensors <- sdb_sensors(types = "GHI")
head(ghi_sensors)

}
\author{
Mathieu Delsaut, \email{mathieu.delsaut@univ-reunion.fr}
}
