% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shiny.R
\name{sdb_shiny_missings}
\alias{sdb_shiny_missings}
\title{Launch SolarDB Missing Data Visualization App}
\usage{
sdb_shiny_missings()
}
\value{
None (invisible). Opens a Shiny application in the default web browser.
}
\description{
Launches a Shiny application in the default web browser that allows users to
visualize and analyze missing data patterns from the SolarDB database.
}
\details{
The application provides an interactive interface for exploring missing data
patterns in SolarDB. If the application cannot be launched due to missing
files, an informative error message will be displayed.
}
\examples{
if (interactive()) {
  sdb_shiny_missings()
}

}
\author{
Mathieu Delsaut, \email{mathieu.delsaut@univ-reunion.fr}
}
