% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\name{sdb_sites}
\alias{sdb_sites}
\title{Get list of alias sites from SolarDB}
\usage{
sdb_sites()
}
\value{
A list of site aliases
}
\description{
Retrieves the list of all available site aliases from the SolarDB API.
}
\examples{
sdb_sites()

}
\author{
Mathieu Delsaut, \email{mathieu.delsaut@univ-reunion.fr}
}
