% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\name{sdb_types}
\alias{sdb_types}
\title{Get list of data types from SolarDB}
\usage{
sdb_types()
}
\value{
A list of data types
}
\description{
Retrieves the list of all available data types from the SolarDB API.
}
\examples{
sdb_types()

}
\author{
Mathieu Delsaut, \email{mathieu.delsaut@univ-reunion.fr}
}
