% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\name{sensors}
\alias{sensors}
\title{Sensors from SolarDB}
\usage{
sensors(sites = NULL, types = NULL)
}
\description{
This function allow know the sensors ids from SolarDB.
}
\examples{
\dontrun{
 sites()
}

}
\author{
Mathieu Delsaut, \email{mathieu.delsaut@univ-reunion.fr}
}
