% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\name{sites}
\alias{sites}
\title{Alias sites from SolarDB}
\usage{
sites()
}
\description{
This function allow know the alias sites from SolarDB.
}
\examples{
\dontrun{
 sites()
}

}
\author{
Mathieu Delsaut, \email{mathieu.delsaut@univ-reunion.fr}
}
