% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deprecated.R
\name{status}
\alias{status}
\title{Status from SolarDB (Deprecated)}
\usage{
status()
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#deprecated}{\figure{lifecycle-deprecated.svg}{options: alt='[Deprecated]'}}}{\strong{[Deprecated]}}

This function is deprecated. Please use \code{sdb_status()} instead.
}
\keyword{internal}
