% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metadata.R
\name{campaigns}
\alias{campaigns}
\title{Campaigns from SolarDB}
\usage{
campaigns(id = NULL, name = NULL, territory = NULL, alias = NULL)
}
\description{
This function allow know the campaigns from SolarDB.
}
\examples{
\dontrun{
 campaigns()
}

}
\author{
Mathieu Delsaut, \email{mathieu.delsaut@univ-reunion.fr}
}
