% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deprecated.R
\name{campaigns}
\alias{campaigns}
\title{Get Campaign Information from SolarDB (Deprecated)}
\usage{
campaigns(id = NULL, name = NULL, territory = NULL, alias = NULL)
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#deprecated}{\figure{lifecycle-deprecated.svg}{options: alt='[Deprecated]'}}}{\strong{[Deprecated]}}

This function is deprecated. Please use \code{sdb_campaigns()} instead.
}
\keyword{internal}
