% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\name{getBounds}
\alias{getBounds}
\title{Get data bounds from SolarDB}
\usage{
getBounds(sites = NULL, types = NULL, sensors = NULL)
}
\description{
This function allow query bounds date of the data from SolarDB.
}
\examples{
\dontrun{
 getBounds()
}

}
\author{
Mathieu Delsaut, \email{mathieu.delsaut@univ-reunion.fr}
}
