% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deprecated.R
\name{getBounds}
\alias{getBounds}
\title{Get data bounds from SolarDB (Deprecated)}
\usage{
getBounds(sites = NULL, types = NULL, sensors = NULL)
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#deprecated}{\figure{lifecycle-deprecated.svg}{options: alt='[Deprecated]'}}}{\strong{[Deprecated]}}

This function is deprecated. Please use \code{sdb_bounds()} instead.
}
\keyword{internal}
