% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deprecated.R
\name{getData}
\alias{getData}
\title{Query time series data from SolarDB (Deprecated)}
\usage{
getData(
  sites = NULL,
  types = NULL,
  sensors = NULL,
  start = NULL,
  stop = NULL,
  aggrFn = NULL,
  aggrEvery = NULL
)
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#deprecated}{\figure{lifecycle-deprecated.svg}{options: alt='[Deprecated]'}}}{\strong{[Deprecated]}}

This function is deprecated. Please use \code{sdb_data()} instead.
}
\keyword{internal}
