% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\name{getData}
\alias{getData}
\title{Get data from SolarDB}
\usage{
getData(
  sites = NULL,
  types = NULL,
  sensors = NULL,
  start = NULL,
  stop = NULL,
  aggrFn = NULL,
  aggrEvery = NULL
)
}
\arguments{
\item{sites}{character (or character vector). List of sites}

\item{types}{character (or character vector). List of types}

\item{sensors}{character (or character vector). List of sensors}

\item{start}{character. Date, Timestamp or duration}

\item{stop}{character. Date, Timestomp or duration}

\item{aggrFn}{character. The aggregation function}

\item{aggrEvery}{character. The aggregation period}
}
\description{
This function allow query the data from SolarDB.
}
\examples{
\dontrun{
 getData()
}

}
\author{
Mathieu Delsaut, \email{mathieu.delsaut@univ-reunion.fr}
}
