% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\name{getDataCountByDay}
\alias{getDataCountByDay}
\title{Get data count by day for a sensor from SolarDB}
\usage{
getDataCountByDay(site, sensor)
}
\description{
This function allow query the data from SolarDB.
}
\examples{
\dontrun{
 getDataCountByDay()
}

}
\author{
Mathieu Delsaut, \email{mathieu.delsaut@univ-reunion.fr}
}
