% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\name{getDtData}
\alias{getDtData}
\title{Get data from SolarDB in data.table format}
\usage{
getDtData(
  sites = NULL,
  types = NULL,
  sensors = NULL,
  start = NULL,
  stop = NULL
)
}
\arguments{
\item{sites}{character (or character vector). List of sites}

\item{types}{character (or character vector). List of types}

\item{sensors}{character (or character vector). List of sensors}

\item{start}{character. Date, Timestamp or duration}

\item{stop}{character. Date, Timestomp or duration}
}
\value{
list of data.table
}
\description{
This function allow query the data from SolarDB and return the result into
the data.table format. The resutl will be a list of data.table by alias site.
}
\examples{
\dontrun{
 getDtData()
}

}
\author{
Mathieu Delsaut, \email{mathieu.delsaut@univ-reunion.fr}
}
