% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\name{getXtsData}
\alias{getXtsData}
\title{Get data from SolarDB into Xts format}
\usage{
getXtsData(
  sites = NULL,
  types = NULL,
  sensors = NULL,
  start = NULL,
  stop = NULL
)
}
\description{
This function allow query the data from SolarDB and return the result into
xts format.
}
\examples{
\dontrun{
 getXtsData()
}

}
\author{
Mathieu Delsaut, \email{mathieu.delsaut@univ-reunion.fr}
}
