% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metadata.R
\name{instruments}
\alias{instruments}
\title{Instruments from SolarDB}
\usage{
instruments(id = NULL, name = NULL, label = NULL, serial = NULL)
}
\description{
This function allow know the instruments from SolarDB.
}
\examples{
\dontrun{
 instruments()
}

}
\author{
Mathieu Delsaut, \email{mathieu.delsaut@univ-reunion.fr}
}
