% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deprecated.R
\name{instruments}
\alias{instruments}
\title{Get Instrument Information from SolarDB (Deprecated)}
\usage{
instruments(id = NULL, name = NULL, label = NULL, serial = NULL)
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#deprecated}{\figure{lifecycle-deprecated.svg}{options: alt='[Deprecated]'}}}{\strong{[Deprecated]}}

This function is deprecated. Please use \code{sdb_instruments()} instead.
}
\keyword{internal}
