% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deprecated.R
\name{login}
\alias{login}
\title{Login to SolarDB with token (Deprecated)}
\usage{
login(token = NULL)
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#deprecated}{\figure{lifecycle-deprecated.svg}{options: alt='[Deprecated]'}}}{\strong{[Deprecated]}}

This function is deprecated. Please use \code{sdb_login()} instead.
}
\keyword{internal}
