% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/login.R
\name{logout}
\alias{logout}
\title{Logout to SolarDB}
\usage{
logout()
}
\description{
This function allow to logout to SolarDB.
}
\examples{
\dontrun{
 logout()
}

}
\author{
Mathieu Delsaut, \email{mathieu.delsaut@univ-reunion.fr}
}
