% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metadata.R
\name{measures}
\alias{measures}
\title{Measures from SolarDB}
\usage{
measures(id = NULL, names = NULL, type = NULL, nested = FALSE)
}
\arguments{
\item{id}{character. The _id from the table measures}

\item{names}{character (or character vector). The names from the table measures}

\item{type}{character. The type from the table measures}

\item{nested}{logical. Answer should be nested from the tables measures, instruments and models}
}
\description{
This function allow know the measures from SolarDB.
}
\examples{
\dontrun{
 measures()
}

}
\author{
Mathieu Delsaut, \email{mathieu.delsaut@univ-reunion.fr}
}
