% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deprecated.R
\name{missings}
\alias{missings}
\title{Launch SolarDB Missing Data Visualization App (Deprecated)}
\usage{
missings()
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#deprecated}{\figure{lifecycle-deprecated.svg}{options: alt='[Deprecated]'}}}{\strong{[Deprecated]}}

This function is deprecated. Please use \code{sdb_shiny_missings()} instead.
}
\keyword{internal}
