% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deprecated.R
\name{models}
\alias{models}
\title{Get Model Information from SolarDB (Deprecated)}
\usage{
models(id = NULL, name = NULL, type = NULL)
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#deprecated}{\figure{lifecycle-deprecated.svg}{options: alt='[Deprecated]'}}}{\strong{[Deprecated]}}

This function is deprecated. Please use \code{sdb_models()} instead.
}
\keyword{internal}
