% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deprecated.R
\name{register}
\alias{register}
\title{Register to SolarDB with an email (Deprecated)}
\usage{
register(email)
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#deprecated}{\figure{lifecycle-deprecated.svg}{options: alt='[Deprecated]'}}}{\strong{[Deprecated]}}

This function is deprecated. Please use \code{sdb_register()} instead.
}
\keyword{internal}
