% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deprecated.R
\name{sensors}
\alias{sensors}
\title{Get list of sensor IDs from SolarDB (Deprecated)}
\usage{
sensors(sites = NULL, types = NULL)
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#deprecated}{\figure{lifecycle-deprecated.svg}{options: alt='[Deprecated]'}}}{\strong{[Deprecated]}}

This function is deprecated. Please use \code{sdb_sensors()} instead.
}
\keyword{internal}
