% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deprecated.R
\name{sites}
\alias{sites}
\title{Get list of alias sites from SolarDB (Deprecated)}
\usage{
sites()
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#deprecated}{\figure{lifecycle-deprecated.svg}{options: alt='[Deprecated]'}}}{\strong{[Deprecated]}}

This function is deprecated. Please use \code{sdb_sites()} instead.
}
\keyword{internal}
