% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/login.R
\name{status}
\alias{status}
\title{Status from SolarDB}
\usage{
status()
}
\description{
This function allow know the login status from SolarDB.
}
\examples{
\dontrun{
 status()
}

}
\author{
Mathieu Delsaut, \email{mathieu.delsaut@univ-reunion.fr}
}
