% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\name{types}
\alias{types}
\title{Types from SolarDB}
\usage{
types()
}
\description{
This function allow know the types from SolarDB.
}
\examples{
\dontrun{
 sites()
}

}
\author{
Mathieu Delsaut, \email{mathieu.delsaut@univ-reunion.fr}
}
